Media_Organizer/requirements.txt
transformers
torch
Pillow
pymediainfo
nudenet
